requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
pyjwt==2.9.0
openai==1.51.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
python-json-logger==2.0.7